backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic[email]==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.12
python-dotenv==1.0.1
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
google-generativeai==0.8.0
